threadcomm: add inter-process pt2pt for threadcomm
Because same process may receive from multiple source tid and dest tid,
we need put both tids into tag for matching. This greatly shrinks the
user usable tag bits.

We are not supporting wild card matching for threadcomm. It will be too
complicated and may not be relevant for the usages of threadcomm.

We modified the tag to carry source id and destination id for
inter-process thread comm messages. We need to filter the tag in filing
the status.