hydra/proxy: split args and procinfo match tables
Split the match tables between parsing the launch args and parsing the
procinfo parameters. The former are global to the proxy, and the latter
are for per-pg.

Also terminate the match tables by NULL-entries. Without it, an
unrecognized option will access over the the matchtables and resulting
in mysterious behaviors.

TODO: move per-pg settings into the pmip_pg struct.